`config`: deprecate `log_compaction_disable_tx_batch_removal`
And add `log_compaction_tx_batch_removal_enabled`.

Unfortunately this previous cluster config shipped early, and with
the wrong default, so to properly gate the behavior of this property
and enabling backport the feature, we have to deprecate the old config
and add a new one which defaults to off (`false`, in the new case).