ct/l0: Share, don't copy, downloaded object when caching
Use iobuf::share not iobuf::copy to obtain an iobuf to read into the
cloud storage cache. This eliminates an extra copy of a fairly hefty
(~4MiB usually) buffer, reducing peak memory usage of the L0 fetch.

Very minor: I moved the creation of the input stream so it happens
after we obtain the space reservation for the object.