`ct/l1`: add `object_metadata_builder::create_object_for()`
In compaction, we need to be able to create individual L1 objects
from the same `object_metadata_builder` for the same partition
concurrently.

Currently, the only interface provided is `get_or_create_object_for()`,
which can result in reuse of the same `object_id`, by design. This has
forced the `compaction_committer` to use a `mutex` around the
`object_metadata_builder` during concurrent constructions of L1 objects to
ensure isolation of objects.

Add `create_object_for()` for this specific use-case, with possible footguns
/unexpected behavior noted in an inline comment above the function declaration.