`ct/l1`: fix `unmanage_cb_t` and some life-time issues
This `unmanage_cb_t` type is typedef'd as returning `void`, when actually
it was invoking a function returning an `ss::future<>`, due to some asynchronous
work that must occur when unmanaging a partition.

Fix the `compaction_scheduler::unmanage_partition()` implementation and some
of the functions in the call-stack so that the asynchronous work is pushed
to a detached fiber lower in `worker_manager`. To ensure proper lifetimes,
add a `ss::gate` to the `worker_manager` and use this to prevent any requests
from entering the `worker_manager` while it is being/after it is stopped.